requirements.txt
arrow==0.12.1
aws-sam-cli==0.4.0
aws-sam-translator==1.6.0
binaryornot==0.4.4
boto3==1.7.50
botocore==1.10.50
certifi==2018.4.16
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
docker==3.4.1
docker-pycreds==0.3.0
docutils==0.14
enum34==1.1.6
feedparser==5.2.1
Flask==1.0.2
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.0
poyo==0.4.1
python-dateutil==2.7.3
PyYAML==3.12
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
websocket-client==0.48.0
Werkzeug==0.14.1
whichcraft==0.4.1
